+++ matlab/schemas/+monet/OriMap.m
@@ -21,7 +21,7 @@ classdef OriMap < dj.Relvar & dj.AutoPopulate
             tic
             disp 'loading design...'
             designMatrix = fetch1(monet.OriDesign & key, 'design_matrix');
-
+                       
             disp 'loading movie...'
             fixRaster = get_fix_raster_fun(pre.AlignRaster & key);
             fixMotion = get_fix_motion_fun(pre.AlignMotion & key);
